Оптимизация (часть 2)
добавил компонент FavoriteEmpty

порефачил Favorite, FavoritePage
