+++ src/components/index.ts
@@ -7,7 +7,7 @@ export * from './city-line/city-line';
 export * from './header/header';
 export * from './footer/footer';
 export * from './offer-list/offer-list';
-export * from './favorite-list/favorite-list';
+export * from './favorite/favorite';
 export * from './review-form/review-form';
 export * from './offer-gallery/offer-gallery';
 export * from './good-list/good-list';
@@ -23,3 +23,4 @@ export * from './review-item/review-item';
 export * from './sorting/sorting';
 export * from './login-form/login-form';
 export * from './bookmark/bookmark';
+export * from './favorite/favorite-empty/favorite-empty';
